TokenHandler is set to DbContext and show in View v0.2
