+++ src/Andoromeda.Kyubey.Models/Hatcher/TokenCommentPraise.cs
@@ -2,12 +2,13 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Andoromeda.Kyubey.Models.TokenHatcher
+namespace Andoromeda.Kyubey.Models.Hatcher
 {
-    public class TokenFollow
+    public class TokenCommentPraise
     {
         public Guid Id { get; set; }
-        public string TokenId { get; set; }
+        public Guid CommentId { get; set; }
+        public bool IsPraise { get; set; }
         public string UserId { get; set; }
         public DateTimeOffset CreateTime { get; set; }
     }
