+++ src/Andoromeda.Kyubey.Models/Hatcher/TokenFollow.cs
@@ -2,13 +2,12 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Andoromeda.Kyubey.Models.TokenHatcher
+namespace Andoromeda.Kyubey.Models.Hatcher
 {
-    public class TokenCommentPraise
+    public class TokenFollow
     {
         public Guid Id { get; set; }
-        public Guid CommentId { get; set; }
-        public bool IsPraise { get; set; }
+        public string TokenId { get; set; }
         public string UserId { get; set; }
         public DateTimeOffset CreateTime { get; set; }
     }
