+++ src/Andoromeda.Kyubey.Models/Hatcher/TokenRecentUpdate.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Andoromeda.Kyubey.Models.TokenHatcher
+namespace Andoromeda.Kyubey.Models.Hatcher
 {
     public class TokenRecentUpdate
     {
